+++ Sources/VComponents/Components/Indicators (Definite)/Page Indicator (Compact)/VCompactPageIndicatorState.swift
@@ -1,0 +1,13 @@
+//
+//  VCompactPageIndicatorState.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 03.09.23.
+//
+
+import Foundation
+import VCore
+
+// MARK: - V Compact Page Indicator Dot Internal State
+/// Enumeration that represents state, such as `deselected` or `selected`.
+public typealias VCompactPageIndicatorDotInternalState = GenericState_DeselectedSelected