+++ Sources/VComponents/Components/Indicators (Definite)/Page Indicator (Compact)/VCompactPageIndicatorUIModel.swift
@@ -16,10 +16,10 @@ public struct VCompactPageIndicatorUIModel {
     public var direction: LayoutDirectionOmni = .leftToRight
 
     /// Dot spacing.
-    /// Set to `5` on `iOS`.
-    /// Set to `5` on `macOS`.
+    /// Set to `8` on `iOS`.
+    /// Set to `8` on `macOS`.
     /// Set to `10` on `tvOS`.
-    /// Set to `3` on `watchOS`.
+    /// Set to `4` on `watchOS`.
     public var spacing: CGFloat = GlobalUIModel.DefiniteIndicators.pageIndicatorSpacing
 
     /// Number of visible dots. Set to `7`.
@@ -32,49 +32,46 @@ public struct VCompactPageIndicatorUIModel {
     /// Must be odd and less than `visibleDots`, otherwise a `fatalError` will occur.
     public var centerDots: Int = GlobalUIModel.DefiniteIndicators.pageIndicatorCompactCenterDots
 
-    /// Number of side dots. Set to half of difference between `visibleDots` and `centerDots`.
+    /// Number of side dots. Set to the half of difference between `visibleDots` and `centerDots`.
     public var sideDots: Int { (visibleDots - centerDots) / 2 }
 
-    /// Number of middle dots. Set to half of `visibleDots`.
+    /// Number of middle dots. Set to the half of `visibleDots`.
     public var middleDots: Int { visibleDots / 2 }
 
     // MARK: Properties - Dot
-    /// Dot width, but height for vertical layouts.
-    /// Set to `10` on `iOS`.
-    /// Set to `10` on `macOS`.
-    /// Set to `20` on `tvOS`.
-    /// Set to `8` on `watchOS`.
+    /// Dot width, but height for vertical layout.
+    /// Set to `8` on `iOS`.
+    /// Set to `8` on `macOS`.
+    /// Set to `10` on `tvOS`.
+    /// Set to `4` on `watchOS`.
     public var dotWidth: CGFloat = GlobalUIModel.DefiniteIndicators.pageIndicatorDotDimension
 
-    /// Dot height, but width for vertical layouts.
-    /// Set to `10` on `iOS`.
-    /// Set to `10` on `macOS`.
-    /// Set to `20` on `tvOS`.
-    /// Set to `8` on `watchOS`.
+    /// Dot height, but width for vertical layout.
+    /// Set to `8` on `iOS`.
+    /// Set to `8` on `macOS`.
+    /// Set to `10` on `tvOS`.
+    /// Set to `4` on `watchOS`.
     public var dotHeight: CGFloat = GlobalUIModel.DefiniteIndicators.pageIndicatorDotDimension
 
     /// Scale of dot at the edge. Set to `0.5`.
     ///
-    /// If there are `7` visible dots, and `3` center dots, scales would sit at `[0.5, 0.75, 1, 1, 1, 0.75, 0.5]`.
+    /// If there are `7` visible dots, and `3` center dots, resulting dot scales would be `[0.5, 0.75, 1, 1, 1, 0.75, 0.5]`.
     public var edgeDotScale: CGFloat = GlobalUIModel.DefiniteIndicators.pageIndicatorCompactEdgeDotScale
 
-    /// Dot color.
-    public var dotColor: Color = GlobalUIModel.DefiniteIndicators.pageIndicatorDotColor
-
-    /// Selected dot color.
-    public var selectedDotColor: Color = GlobalUIModel.DefiniteIndicators.pageIndicatorSelectedDotColor
+    /// Dot colors.
+    public var dotColors: DotStateColors = .init(
+        deselected: GlobalUIModel.DefiniteIndicators.pageIndicatorDeselectedDotColor,
+        selected: GlobalUIModel.DefiniteIndicators.pageIndicatorSelectedDotColor
+    )
 
     // MARK: Properties - Dot Border
-    /// Border width. Set to `0`.
+    /// Dot border widths. Set to `0`s.
     ///
-    /// To hide border, set to `0`.
-    public var dotBorderWidth: CGFloat = 0
+    /// To hide border, set to `0`s.
+    public var dotBorderWidths: DotStateDimensions = .zero
 
-    /// Dot border color.
-    public var dotBorderColor: Color = .clear
-
-    /// Selected dot border color.
-    public var selectedDotBorderColor: Color = .clear
+    /// Dot border colors.
+    public var dotBorderColors: DotStateColors = .clearColors
 
     // MARK: Properties - Transition
     /// Indicates if `transition` animation is applied. Set to `true`.
@@ -91,30 +88,15 @@ public struct VCompactPageIndicatorUIModel {
     public var transitionAnimation: Animation? = GlobalUIModel.DefiniteIndicators.pageIndicatorTransitionAnimation
 
     // MARK: Properties - Standard Layout
-    /// Dot width, but height for vertical layouts, when switching to `standard` configuration.
-    /// Set to `10` on `iOS`.
-    /// Set to `10` on `macOS`.
-    /// Set to `20` on `tvOS`.
-    /// Set to `8` on `watchOS`.
-    ///
-    /// Set to `nil`, to make dot stretch to take available space.
-    public var dotWidthForStandardConfiguration: CGFloat? = GlobalUIModel.DefiniteIndicators.pageIndicatorDotDimension
-
-    /// Unselected dot scale when switching to `standard` configuration. Set to `0.85`.
-    public var unselectedDotScaleForStandardConfiguration: CGFloat = GlobalUIModel.DefiniteIndicators.pageIndicatorStandardUnselectedDotScale
-
     var standardPageIndicatorSubUIModel: VPageIndicatorUIModel {
         .init(
             direction: direction,
             spacing: spacing,
-            dotWidth: dotWidthForStandardConfiguration,
-            dotHeight: dotHeight,
-            unselectedDotScale: unselectedDotScaleForStandardConfiguration,
-            dotColor: dotColor,
-            selectedDotColor: selectedDotColor,
-            dotBorderWidth: dotBorderWidth,
-            dotBorderColor: dotBorderColor,
-            selectedDotBorderColor: selectedDotBorderColor,
+            dotWidths: VPageIndicatorUIModel.DotStateOptionalDimensions(dotWidth),
+            dotHeights: VPageIndicatorUIModel.DotStateDimensions(dotHeight),
+            dotColors: dotColors,
+            dotBorderWidths: dotBorderWidths,
+            dotBorderColors: dotBorderColors,
             appliesTransitionAnimation: appliesTransitionAnimation,
             transitionAnimation: transitionAnimation
         )
@@ -123,6 +105,14 @@ public struct VCompactPageIndicatorUIModel {
     // MARK: Initializers
     /// Initializes UI model with default values.
     public init() {}
+
+    // MARK: Dot State Dimensions
+    /// Model that contains dimensions  for component states.
+    public typealias DotStateDimensions = GenericStateModel_DeselectedSelected<CGFloat>
+
+    // MARK: Dot State Colors
+    /// Model that contains colors for component states.
+    public typealias DotStateColors = GenericStateModel_DeselectedSelected<Color>
 }
 
 // MARK: - Factory
@@ -130,9 +120,9 @@ extension VCompactPageIndicatorUIModel {
     /// `VCompactPageIndicatorUIModel` with vertical layout.
     public static var vertical: Self {
         var uiModel: Self = .init()
-        
+
         uiModel.direction = .topToBottom
-        
+
         return uiModel
     }
 }