+++ Sources/VComponents/Components/Indicators (Definite)/Page Indicator/VPageIndicatorState.swift
@@ -1,0 +1,13 @@
+//
+//  VPageIndicatorState.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 03.09.23.
+//
+
+import Foundation
+import VCore
+
+// MARK: - V Page Indicator Dot Internal State
+/// Enumeration that represents state, such as `deselected` or `selected`.
+public typealias VPageIndicatorDotInternalState = GenericState_DeselectedSelected