+++ Sources/VComponents/Components/Inputs/Code Entry View/VCodeEntryView.swift
@@ -37,8 +37,7 @@ import VCore
 ///
 /// Highlights can be applied using `success`, `warning`, and `secure` instances of `VCodeEntryViewUIModel`.
 ///
-/// Component may cause issues when rendered in `SwiftUI` previews.
-/// When running is simulators, `mac` keyboard may cause the hidden `TextField` decoration to appear.
+/// When running in `SwiftUI` previews or simulators, `mac` keyboard may cause the hidden `TextField` decoration to appear.
 @available(iOS 15.0, *)
 @available(macOS 12.0, *)@available(macOS, unavailable) // Doesn't follow Human Interface Guidelines
 @available(tvOS 16.0, *)@available(tvOS, unavailable) // Doesn't follow Human Interface Guidelines